Create users route & connect to postgres container

+++ src/models/user.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(sqlx::FromRow, Deserialize, Serialize)]
 pub struct User {
-    id: i64,
+    id: i32,
     username: String,
     password: String
 }
